+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## changelog
 
+
+### v0.11.5
+- changed, _vars.php renamed for ../cred.{domain}_vars.php
+- changed, NROK tunnel = htdocs
+
 ### v0.11.4
 - fixed, index had old links to kista-ai.steinhaug.no, now set in _vars.php
 - fixed, Google API has updated syntax for oauth login.